Initial brute force version of the algorithm

+++ src/main/java/com/satalia/beer/dto/CalculatedPathResponse.java
@@ -18,6 +18,6 @@ public class CalculatedPathResponse {
 
     private List<String> breweriesVisited;
     private List<String> beersCollected;
-    private int distanceTraveled;
+    private Double distanceTraveled;
 
 }
